Server fully starts now

+++ Source/ACRealms.Rulesets/DBOld/RealmPropertiesBase.cs
@@ -10,7 +10,7 @@ namespace ACRealms.Rulesets.DBOld
     public abstract class RealmPropertiesBase
     {
         public ushort RealmId { get; set; }
-        public ushort Type { get; set; }
+        public int Type { get; set; }
         public bool Locked { get; set; }
         public double? Probability { get; set; }
         public virtual Realm Realm { get; set; }
